Restructuring and refactoring to allow for scalability when adding new API wrappers/parsers.
Moved Base from products.base into shopify.base and renamed to BaseParser so that when working on the orders section of the api I dont have to re-write a lot of code which makes
the library more maintainable and scalable.

The methods convert_timestamp and convert_datetime are no longer a classmethod to BaseParser but instead are a function of the base.py module.